LayoutTests/fast/tokenizer/preload-scanner-entities.html

<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.dumpResourceLoadCallbacks();
}
</script>
<script src="does_not_exist_here_just_to_trigger_preload_scanner.js" type="text/javascript"></script>
<link href="non_existent_stylesheet1&ab-" rel="stylesheet">
<link href="non_existent_stylesheet2&#xY" rel="stylesheet">
<link href="non_existent_stylesheet3&#XY" rel="stylesheet">
<p>If you run this test outside the DRT, check with the web inspector that the page tried to load 3 stylesheet resources (and not more).</p>
